V26-749 refactor register view model boundaries
