+++ packages/athena-webapp/docs/agent/test-index.md
@@ -322,6 +322,12 @@ This index enumerates the current automated test files and ties them back to the
 - [`src/lib/pos/presentation/expense/useExpenseRegisterViewModel.test.ts`](../../src/lib/pos/presentation/expense/useExpenseRegisterViewModel.test.ts)
 - [`src/lib/pos/presentation/register/catalogSearch.test.ts`](../../src/lib/pos/presentation/register/catalogSearch.test.ts)
 - [`src/lib/pos/presentation/register/catalogSearchPresentation.test.ts`](../../src/lib/pos/presentation/register/catalogSearchPresentation.test.ts)
+- [`src/lib/pos/presentation/register/registerCartProjection.test.ts`](../../src/lib/pos/presentation/register/registerCartProjection.test.ts)
+- [`src/lib/pos/presentation/register/registerCashierPresence.test.ts`](../../src/lib/pos/presentation/register/registerCashierPresence.test.ts)
+- [`src/lib/pos/presentation/register/registerCheckoutProjection.test.ts`](../../src/lib/pos/presentation/register/registerCheckoutProjection.test.ts)
+- [`src/lib/pos/presentation/register/registerDrawerPresentation.test.ts`](../../src/lib/pos/presentation/register/registerDrawerPresentation.test.ts)
+- [`src/lib/pos/presentation/register/useRegisterCheckoutDraftState.test.ts`](../../src/lib/pos/presentation/register/useRegisterCheckoutDraftState.test.ts)
+- [`src/lib/pos/presentation/register/useRegisterLocalRuntime.test.ts`](../../src/lib/pos/presentation/register/useRegisterLocalRuntime.test.ts)
 - [`src/lib/pos/presentation/register/useRegisterViewModel.test.ts`](../../src/lib/pos/presentation/register/useRegisterViewModel.test.ts)
 - [`src/lib/pos/presentation/syncStatusPresentation.test.ts`](../../src/lib/pos/presentation/syncStatusPresentation.test.ts)
 - [`src/lib/productUtils.test.ts`](../../src/lib/productUtils.test.ts)
